+++ Sources/BraintreeShopperInsights/BTShopperInsightsError.swift
@@ -4,7 +4,10 @@ public enum BTShopperInsightsError: Int, Error, CustomNSError, LocalizedError, E
     
     /// 0. A nil body was returned from the payment method request and no error was returned.
     case emptyBodyReturned
-    
+
+    /// 1. Invalid authorization type
+    case invalidAuthorization
+
     public static var errorDomain: String {
         "com.braintreepayments.BTShopperInsightsErrorDomain"
     }
@@ -17,6 +20,8 @@ public enum BTShopperInsightsError: Int, Error, CustomNSError, LocalizedError, E
         switch self {
         case .emptyBodyReturned:
             return "An empty body was returned from the Eligible Payments API during the request."
+        case .invalidAuthorization:
+            return "Invalid authorization. This feature can only be used with a client token."
         }
     }
 }