+++ UnitTests/BraintreeShopperInsightsTests/BTShopperInsightsClient_Tests.swift
@@ -6,7 +6,8 @@ import XCTest
 
 class BTShopperInsightsClient_Tests: XCTestCase {
     
-    var mockAPIClient = MockAPIClient(authorization: "development_client_key")!
+    let clientToken = TestClientTokenFactory.token(withVersion: 3)
+    var mockAPIClient: MockAPIClient!
     var sut: BTShopperInsightsClient!
     
     let request = BTShopperInsightsRequest(
@@ -19,7 +20,8 @@ class BTShopperInsightsClient_Tests: XCTestCase {
     
     override func setUp() {
         super.setUp()
-        sut = BTShopperInsightsClient(apiClient: mockAPIClient)
+        mockAPIClient = MockAPIClient(authorization: clientToken)
+        sut = BTShopperInsightsClient(apiClient: mockAPIClient!)
     }
     
     // MARK: - getRecommendedPaymentMethods()
@@ -30,7 +32,7 @@ class BTShopperInsightsClient_Tests: XCTestCase {
         XCTAssertEqual(mockAPIClient.lastPOSTPath, "/v2/payments/find-eligible-methods")
         XCTAssertEqual(mockAPIClient.lastPOSTAPIClientHTTPType, .payPalAPI)
         XCTAssertEqual(mockAPIClient.lastPOSTAdditionalHeaders?["PayPal-Client-Metadata-Id"], mockAPIClient.metadata.sessionID)
-        
+
         guard let lastPostParameters = mockAPIClient.lastPOSTParameters else {
             XCTFail()
             return
@@ -55,7 +57,7 @@ class BTShopperInsightsClient_Tests: XCTestCase {
     
     func testGetRecommendedPaymentMethods_whenAPIError_throws() async {
         mockAPIClient.cannedResponseError = NSError(domain: "fake-error-domain", code: 123, userInfo: [NSLocalizedDescriptionKey:"fake-error-description"])
-        
+
         do {
             _ = try await sut.getRecommendedPaymentMethods(request: request)
             XCTFail("Expected error to be thrown.")
@@ -171,7 +173,20 @@ class BTShopperInsightsClient_Tests: XCTestCase {
             XCTFail("An error was not expected.")
         }
     }
-    
+
+    func testGetRecommendedPaymentMethods_withTokenizationKey_returnsError() async {
+        var apiClient = BTAPIClient(authorization: "sandbox_merchant_1234567890abc")!
+        let shopperInsightsClient = BTShopperInsightsClient(apiClient: apiClient)
+
+        do {
+            let result = try await shopperInsightsClient.getRecommendedPaymentMethods(request: request)
+        } catch {
+            let error = error as NSError
+            XCTAssertEqual(error.code, 1)
+            XCTAssertEqual(error.localizedDescription, "Invalid authorization. This feature can only be used with a client token.")
+        }
+    }
+
     // MARK: - Analytics
     
     func testSendPayPalPresentedEvent_sendsAnalytic() {